edit : roomMate Search 삭제, PostCard 수정
